chore: Reorganize graphql queries types in bff, move mutations to its own file
