+++ packages/bff/src/graphql/types/index.ts
@@ -2,3 +2,4 @@ export * from './profile.ts';
 export * from './query.ts';
 export * from './savedSearches.ts';
 export * from './organization.ts';
+export * from './mutation.ts';
