+++ packages/bff/src/graphql/types/profile.ts
@@ -5,13 +5,13 @@ export const Profile = objectType({
   definition(t) {
     t.string('language', {
       description: 'Preferred language for the profile',
-      resolve: (profile, args, ctx, info) => {
+      resolve: (profile) => {
         return profile.language;
       },
     });
     t.string('updatedAt', {
       description: 'Last updated',
-      resolve: (profile, args, ctx, info) => {
+      resolve: (profile) => {
         return profile.updatedAt;
       },
     });
